#218 [refactor] 지역 리스트 조회 Controller 이동
